+++ R/enrich_PathfindR.R
@@ -80,8 +80,6 @@ enrich_pathfindR <- function(study,
           area  = "GENE",
           scope = "INSTANCE",
           marker = keys[i,"MARKER"],
-          adjust_per_area= adjust_per_area,
-          adjust_globally = adjust_globally,
           pvalue_column=  pvalue_column,
           adjustment_method= adjustment_method,
           significance = TRUE)